requirements.txt
streamlit>=1.28.0
langchain-google-genai>=0.0.6
langchain-core>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
plotly>=5.17.0
pandas>=2.0.0
reportlab>=4.0.0
fpdf2>=2.7.0
fastapi>=0.100.0
uvicorn>=0.23.0
